requirements.txt
PyHamcrest
nose
